refactor: move deck start filters to page owner (#899)
