--- src/features/deck-start/index.ts
@@ -1,2 +1,0 @@
-export { DeckStartForm } from "./ui/DeckStartForm";
-export { useDeckFilterState } from "./model/useDeckFilterState";
