+++ src/pages/deck-start/ui/DeckStartForm.tsx
@@ -1,5 +1,5 @@
 /**
- * @file Defines the deck-start feature's Deck Start Form presentation component.
+ * @file Defines the Deck Start page's filter form presentation component.
  * The component renders props and reports user intent through callbacks while data access stays
  * outside the view.
  */
